QL_Nhan_Su_Python/requirements.txt
# ===== Các thư viện cần thiết cho đồ án Quản Lý Nhân Sự =====

# Kết nối Python với MySQL
mysql-connector-python==9.1.0

# Đọc file .env để cấu hình database
python-dotenv==1.1.1

# Lịch chọn ngày sinh trong Tkinter
tkcalendar==1.6.1

# Hỗ trợ xử lý và hiển thị hình ảnh (icon, avatar)
Pillow==11.0.0

# Tùy chọn: giúp xuất dữ liệu hoặc báo cáo (nếu mở rộng tính năng)
openpyxl==3.1.5
reportlab==4.2.2
